app/requirements.txt
asyncio
azure-iot-device
azure.iot.hub